==> 1 - Tools Overview.ipynb <==
<hr>
# PYTHON FOR DATA SCIENCE

- Part 1: Overview
- Part 2: Case Study

### An Overview of Scientific Python

## Contents

- [IPython](#IPython)
- [NumPy](#NumPy)
- [SciPy](#SciPy)
- [Matplotlib](#Matplotlib)
- [pandas](#pandas)
- [scikit-learn](#scikit-learn)
- [statsmodels](#statsmodels)

<a id="IPython"></a>
<hr>

## IPython

Command line and web IDE designed for use with Python and several other scientific and web-based technologies.

#### What is it?

### A computing architecture for Python composed of 

 - 2 Python shells (terminal and Qt-based)

- A web notebook that runs in the browser.

### What can you do with it?

## Code completion

#### ```IPython``` tells you about the objects you're working with, showing you helpful usage info about functions, methods available, datatypes, and more.

things = ["First Name", "Last Name"]

def snakify(txt):
    "return string in snake_case"
    return txt.replace(" ","_").lower()

print [snakify(thing) for thing in things]


snakify( "My Name is Greg" )

<bold class="bold">Tip: </bold>Comment and uncomment lines with ``` command + / ``` if you're on a mac.

## Interact with the command-line

#### Use many shell commands directly from IPython, including shell scripts.

# try these out too
! pwd
!ls -lt

#### Capture the output of a shell command and store the results in a python variable.

foo = ! echo hello, world
foo

#### Use curl to download data.

# download salary_data.csv save contents to 
# a local file in the same directory as this notebook
!curl http://www.justinmrao.com/salary_data.csv >> ./salary_data.csv

# peak at the first row
! head -n 1 salary_data.csv | tr -s "," "\n"

# peak at the second row
! head -n+2 salary_data.csv | tail -n-1 | tr -s "," "\n"

## Include rich media and javascript
#### IPython gives you inline plots and allows you to use virtually any browser supported media.

Inline plotting

import pylab as pl
X = range(10)
y = range(11,21)
pl.scatter(X,y, c='r')
pl.show()
print

#### Write and run javascript among other languages including Ruby, Octave, Julia, R, etc.

%%javascript

function say_hi ( ) {
    alert("Hello, World");
};

say_hi();
console.log("Welcome!");

#### Render images from the web via a url or images on your local computer via the filename.

from IPython.display import Image
Image("http://ipython.org/_static/IPy_header.png")

<a id="NumPy"></a>
<hr>

## NumPy

One of the fundamental packages for scientific computing with Python. Many other popular packages and projects use NumPy under the hood, so it's pretty helpful to be familiar with its core concepts.

#### What is it?

### A collection of tools, utilities, and data structures for working with data.

- A powerful N-dimensional array object
- Element by element broadcasting operations (as opposed to iterating)
- Tools for integrating C/C++ and Fortran code
- Linear algebra and other mathematical and random number facilities



## N-dimensional arrays

#### The ```ndarray``` is a bit like a python list only more efficient.

"""example taken from Scipy and NumPy by Eli Bressert (p. 6)"""
import numpy as np

def list_times(alist, scalar):
    for i, val in enumerate(alist):
        alist[i] = val * scalar
    return alist

arr = np.arange(1e7)
l   = arr.tolist()

%timeit arr * 1.1

%timeit list_times(l, 1.1)

#### ```ndarray```s behave somewhat like lists.

print "len(l)", len(l)
print "len(arr)", len(arr)

#### The ```ndarray``` is capible of more advanced slicing.

l = [ [1,2], [3,4] ]
arr   = np.array(l)

print "Value in Row One, Column One: %d" % l[0][0]
print "Value in Row One, Column One: %d" % arr[0,0]
print "Value in All Rows, Column Two: %s" % str(arr[::,1])
print "Value in Row Two, Both Columns: %s" % str(arr[1::,])

#### You can access elements stored in an ```ndarray``` using ```arr[]``` sub notation ```arr[rows, columns]```.

#### Select the entire 10th row like so ```arr[9, :]```.  ```9``` indicates "row 10" and ```:``` indicates <em>all</em> columns.

zero_to_1000 = np.arange(0,1000)               # create an array of integers from 0 to 1000
zero_to_1000 = zero_to_1000.reshape( (500,2) ) # reshape into 2 dimensions (500 Rows x 2 Cols2)
zero_to_1000[:100, 1]                          # select the 2nd columns from the top 100 rows

#### NumPy arrays support boolean indexing for selection.

numbers = np.random.uniform(size=100)
numbers = numbers.reshape((50,2))
mask = (numbers >= 0.7) & (numbers <= 0.9)
mask[:10]

numbers[mask, :] # pass the boolean mask to the array using sub notation

<a id="SciPy"></a>
<hr>

## SciPy

SciPy contains functions and utilities for common scientific tasks. Most of SciPy's features rely on NumPy and many other scientific projects rely on SciPy.


We aren't using SciPy directly in this tutorial, and since it's rather large in scope, we'll just give you some of the project highlights.

- Optimization and Minimization
    - 80 continuous distributions
    - 60 statistical functions
- Interpolation
- Integration
- Statistics
- Spatial and Clustering Analysis
- Signal and Image Processing
- Sparse Matrices

<a id="Matplotlib"></a>
<hr>

## Matplotlib

import pylab as pl
import matplotlib.pyplot as plt

# Make some data to plot
x = np.linspace(start=0, stop=2*np.pi, num=50)
y1 = np.sin(x)
y2 = np.cos(x)

#### Lines

fig, ax = plt.subplots(1, figsize=(12,6))
ax.plot(x, y1, label='sin')
ax.plot(x, y2, label='cos')

ax.legend(fontsize=16)
ax.axes.xaxis.set_label_text("X Axis", fontdict={"size":22})
ax.axes.yaxis.set_label_text("y Axis", fontdict={"size":22})
ax.title.set_text("This is a graph\n")
ax.title.set_fontsize(28)
fig.show()
print

#### Histograms

d = randn(100) * 100.
m = d.mean()
s = d.std()
m_y = 1.5

fig = figure(figsize=(12,6))
ax = subplot(111)
ax.hist(d, 15)
ax.plot(m, m_y, "ko")
ax.plot([m - s, m + s], [m_y] * 2, "k--");

#### Multiple plots.

x = np.arange(0, 100)
y = np.random.rand(100)  # 100 random numbers
fig, ((ax1, ax2), (ax3, ax4)) = plt.subplots(2,2,figsize=(10,10))
ax1.plot(x, y)
ax2.hist(y)
ax3.scatter(x, y)
ax4.boxplot(y)
pl.show()

## ggplot2 matplotlib theme

- [https://s3.amazonaws.com/demo-datasets/matplotlib.zip](https://s3.amazonaws.com/demo-datasets/matplotlib.zip)

#if you already have a matplotlib config file, you'll need to run this in the terminal
! unzip ~/Downloads/matplotlib.zip -d ~/

#### Matplotlib takes some getting used to, but don't fret. There are some solid examples and recipes available online that are super helpful.

[http://matplotlib.org/gallery.html](http://matplotlib.org/gallery.html.)

<a id="pandas"></a>
<hr>

## pandas

#### pandas is the utility belt for data analysts using python. The package centers around the pandas ```DataFrame```, a two-dimensional data structure with indexable rows and columns. It has effectively taken the best parts of Base R, R packages like plyr and reshape2 and consolidated them into a single library.

#### Library highlights include:

- DataFrame
- Indexable rows and columns (e.g. selecting columns and rows via column/row names)
- Boolean indexing
- Graceful treatment of missing or ```NA``` values
- Broadcasting (i.e. calling a function on an entire row or column without iteration)
- Outstanding ```split```, ```apply```, ```combine```, ```groupby``` functionality
- Simple yet powerful string manipulation (again, applied over a large number of strings)
- Built-in summary statistics (```mean```, ```std```, ```corr```, etc.)
- Merging, joining, data set concatenation


import pandas as pd
df = pd.DataFrame({"A": range(10), "B": np.random.random(size=10)})
df.B.corr(df.A)

<a id="scikit-learn"></a>
<hr>

## scikit-learn

scikit-learn is composed of a wide number of machine learning algorithms, utilities, and transformations all following a common and consistent API.

#### Library highlights include:

- Biclustering
- Clustering
- Covariance estimation
- Cross decomposition
- Dataset examples
- Decomposition
- Ensemble methods
- Gaussian Process for Machine Learning
- Generalized Linear Models
- Manifold learning
- Gaussian Mixture Models
- Nearest Neighbors
- Semi Supervised Classification
- Support Vector Machines
- Decision Trees

%load http://scikit-learn.org/stable/_downloads/plot_lena_ward_segmentation.py

"""
===============================================================
A demo of structured Ward hierarchical clustering on Lena image
===============================================================

Compute the segmentation of a 2D image with Ward hierarchical
clustering. The clustering is spatially constrained in order
for each segmented region to be in one piece.
"""

# Author : Vincent Michel, 2010
#          Alexandre Gramfort, 2011
# License: BSD 3 clause

print(__doc__)

import time as time
import numpy as np
import scipy as sp
import pylab as pl
from sklearn.feature_extraction.image import grid_to_graph
from sklearn.cluster import Ward

###############################################################################
# Generate data
lena = sp.misc.lena()
# Downsample the image by a factor of 4
lena = lena[::2, ::2] + lena[1::2, ::2] + lena[::2, 1::2] + lena[1::2, 1::2]
X = np.reshape(lena, (-1, 1))

###############################################################################
# Define the structure A of the data. Pixels connected to their neighbors.
connectivity = grid_to_graph(*lena.shape)

###############################################################################
# Compute clustering
print("Compute structured hierarchical clustering...")
st = time.time()
n_clusters = 15  # number of regions
ward = Ward(n_clusters=n_clusters, connectivity=connectivity).fit(X)
label = np.reshape(ward.labels_, lena.shape)
print("Elapsed time: ", time.time() - st)
print("Number of pixels: ", label.size)
print("Number of clusters: ", np.unique(label).size)

###############################################################################
# Plot the results on an image
pl.figure(figsize=(5, 5))
pl.imshow(lena, cmap=pl.cm.gray)
for l in range(n_clusters):
    pl.contour(label == l, contours=1,
               colors=[pl.cm.spectral(l / float(n_clusters)), ])
pl.xticks(())
pl.yticks(())
pl.show()


<a id="statsmodels"></a>
<hr>

## statsmodels
####statistics & econometrics package with useful tools for parameter estimation & statistical testing

####Features include:
- linear regression models
- GLMs
- time series modeling
- integration with `pandas`

%load https://gist.github.com/glamp/6536064/raw/57146ac3406f7541d944264ed9aa85dad6e085a1/statsmodels_example.py

import pandas as pd
import statsmodels.api as sm
import pylab as pl
import numpy as np
 
# read the data in
df = pd.read_csv("http://www.ats.ucla.edu/stat/data/binary.csv")
 
# rename the 'rank' column because there is also a DataFrame method called 'rank'
df.columns = ["admit", "gre", "gpa", "prestige"]
dummy_ranks = pd.get_dummies(df['prestige'], prefix='prestige')
dummy_ranks.head()

# create a clean data frame for the regression
cols_to_keep = ['admit', 'gre', 'gpa']
data = df[cols_to_keep].join(dummy_ranks.ix[:, 'prestige_2':])
data.head()

# manually add the intercept
data['intercept'] = 1.0

train_cols = data.columns[1:]
# Index([gre, gpa, prestige_2, prestige_3, prestige_4], dtype=object)
 
logit = sm.Logit(data['admit'], data[train_cols])
 
# fit the model
result = logit.fit()

print result.summary()
