==> 3 - scikit-learn basics.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- <p style="color: red">Intro to scikit-learn</p>
- Grouping and aggregating data
- Feature selection
- Fitting and evaluating a model
- Deploying your work

##In this notebook you will

- Take a tour of `scikit-learn` and learn what it's used for
- Build a toy classifier
- Make and vizualize a decision tree
- Build your own regression model

import pandas as pd
import numpy as np
import pylab as pl

##[scikit-learn](http://scikit-learn.org/stable/)

####Consistent APIs
Algorithms are implemented with the same core functions:

- fit = train an algorithm
- predict = predict the value for a given record
- predict_proba = predict the probability of all possible classes for a given record (classification only)
- transform = alter your data based on a given preprocessor (i.e. normalize or scale your data) (preprocessing/unsuperivsed)
- fit_transform = train a preprocessor and then transform the data in a single step (preprocessing/unsuperivsed)

<blockquote class="twitter-tweet"><p>Scikit-Learn reply to today&#39;s <a href="https://twitter.com/wiseio">@wiseio</a> Random Forest benchmark: <a href="https://t.co/El5at9KvHS">https://t.co/El5at9KvHS</a> … Coming soon in the next 0.14 stable release!</p>&mdash; Gilles Louppe (@glouppe) <a href="https://twitter.com/glouppe/statuses/357250042538622976">July 16, 2013</a></blockquote>
<script async src="//platform.twitter.com/widgets.js" charset="utf-8"></script>

from sklearn.datasets import load_iris

iris = load_iris()
df = pd.DataFrame(iris.data, columns=iris.feature_names)
df['species'] = iris.target

from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier

svm_clf = SVC()
neighbors_clf = KNeighborsClassifier()
clfs = [
    ("svc", SVC()),
    ("KNN", KNeighborsClassifier())
    ]
for name, clf in clfs:
    clf.fit(df[iris.feature_names], df.species)
    print name, clf.predict(iris.data)
    print "*"*80

####Try a [RandomForetClassifier](http://scikit-learn.org/stable/modules/ensemble.html#random-forests) (`from sklearn.ensemble import RandomForestClassifier`) and see how it does

from sklearn.ensemble import RandomForestClassifier
clf = RandomForestClassifier()
clf.fit(df[iris.feature_names], df.species)
clf.predict(df[iris.feature_names])
pd.crosstab(df.species, clf.predict(df[iris.feature_names]))

##A Quick Decision Tree Example
####[more examples](http://scikit-learn.org/stable/auto_examples/index.html#decision-trees)

####Import the decision tree library and train a classifier

from sklearn import tree

clf = tree.DecisionTreeClassifier(max_features="auto",
                                  min_samples_leaf=10)
clf.fit(df[iris.feature_names], df.species)

####Write the tree to a file

from sklearn.externals.six import StringIO
with open("iris.dot", 'w') as f:
    f = tree.export_graphviz(clf, out_file=f)

####We're going to convert it to a .png so we can display it in the notebook

# you will need to install graphviz 
#(http://www.graphviz.org/Download..php) and pydot (pip install pydot)
! dot -Tpng iris.dot -o iris.png

from IPython.core.display import Image
Image("iris.png")

##What to do first?
[Andy Mueller](http://peekaboo-vision.blogspot.com/) (scikit-learn contributor) put together this [cheat sheet](http://1.bp.blogspot.com/-ME24ePzpzIM/UQLWTwurfXI/AAAAAAAAANw/W3EETIroA80/s1600/drop_shadows_background.png) a few months ago which is extremely helpful.

Image(url="http://1.bp.blogspot.com/-ME24ePzpzIM/UQLWTwurfXI/AAAAAAAAANw/W3EETIroA80/s1600/drop_shadows_background.png",
      width=700)

##Try building a linear regression model for the Boston Housing Dataset to predict Home Prices

from sklearn.datasets import load_boston
boston = load_boston()

##Prep the data

- Create a dataframe with the Boston data
- snake_case/lower_case the columns
- print to the console

import re


def camel_to_snake(column_name):
    """
    converts a string that is camelCase into snake_case
    Example:
        print camel_to_snake("javaLovesCamelCase")
        > java_loves_camel_case
    See Also:
        http://stackoverflow.com/questions/1175208/elegant-python-function-to-convert-camelcase-to-camel-case
    """
    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', column_name)
    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()

df = pd.DataFrame(boston.data)
df.columns = [camel_to_snake(col) for col in boston.feature_names[:-1]]
# add in prices
df['price'] = boston.target
print len(df)==506
df.head()

##Fit your model

- define your features
- create a LinearRegression
- fit the model

from sklearn.linear_model import LinearRegression

features = ['age', 'lstat', 'tax']
lm = LinearRegression()
lm.fit(df[features], df.price)

##Plot the predicted values against the actual values

# add your actual vs. predicted points
pl.scatter(df.price, lm.predict(df[features]))
# add the line of perfect fit
straight_line = np.arange(0, 60)
pl.plot(straight_line, straight_line)
pl.title("Fitted Values")

###Redo the model, but this time have a training/test set.

##We just did the following

- Learned about `scikit-learn` and what it's used for
- Used CART to build a decision tree classifier and visualized it
- Built a regression model using `scikit-learn`
