==> 6 - Feature Engineering.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- Intro to scikit-learn
- Grouping and aggregating data
- <p style="color: red">Feature selection</p>
- Fitting and evaluating a model
- Deploying your work

#In this workbook you will
- Use RandomForest to calculate the best features for your model
- Handling outlying data using `apply` and `groupby`
- Discretize (or bucket) data into groups
- Make your own features for your model

import pandas as pd
import numpy as np
import pylab as pl

df = pd.read_csv("./data/credit-data-trainingset.csv")

df.head()

##Finding Important Features
We're going to let scikit-learn help us determine which variables are the best at predicting risk. To do this, we're going to use an algorithm called `RandomForest`. `RandomForest` randomly generates a "forest" of decision trees. As the trees are randomly generated, the algorithm takes turns leaving out each variable in fitting the model. This allows the `RandomForest` to calculate just how much worse a model does when each variable is left out.

from sklearn.ensemble import RandomForestClassifier

features = np.array(['revolving_utilization_of_unsecured_lines',
                     'age', 'number_of_time30-59_days_past_due_not_worse',
                     'debt_ratio', 'monthly_income','number_of_open_credit_lines_and_loans', 
                     'number_of_times90_days_late', 'number_real_estate_loans_or_lines',
                     'number_of_time60-89_days_past_due_not_worse', 'number_of_dependents'])

clf = RandomForestClassifier(compute_importances=True)
clf.fit(df[features], df['serious_dlqin2yrs'])

# from the calculated importances, order them from most to least important
# and make a barplot so we can visualize what is/isn't important
importances = clf.feature_importances_
sorted_idx = np.argsort(importances)

padding = np.arange(len(features)) + 0.5
pl.barh(padding, importances[sorted_idx], align='center')
pl.yticks(padding, features[sorted_idx])
pl.xlabel("Relative Importance")
pl.title("Variable Importance")
pl.show()

So you can see that the best variable is `revolving_utilization_of_unsecured_lines` while the worst is `number_real_estate_loans_or_lines`. There's also a dramatic drop off after `number_of_open_credit_lines_and_loans`. This is where you need to use your own discretion. How many variables should you include in your model?

###Engineering Additional Features
Feature selection/engineering will likely have the biggest impact in determining the success/failure of your model. Even if you're using the latest and greatest algorithm, if you put in non-important features you're going to get poor results. Remember, it's math, not magic.

Feature engineering is a skill that will take time to get the hang of. Sometimes the best way is to just talk to people. Ask questions, brainstorm with others, etc. Oftentimes 2 features might not be helpful when used individually, but when combined they can be extremely powerful.



df['income_bins'] = pd.cut(df.monthly_income, bins=15)
pd.value_counts(df['income_bins'])
# not very helpful

###Bucketing Continuous Values

####Use the cap_value function you wrote previously to cap monthly_income at $15,000

def cap_values(x, cap):
    if x > cap:
        return cap
    else:
        return x
    
df.monthly_income = df.monthly_income.apply(lambda x: cap_values(x, 15000))

df.monthly_income.describe()

df['income_bins'] = pd.cut(df.monthly_income, bins=15, labels=False)
pd.value_counts(df.income_bins)

df[["income_bins", "serious_dlqin2yrs"]].groupby("income_bins").mean()

cols = ["income_bins", "serious_dlqin2yrs"]
df[cols].groupby("income_bins").mean().plot()

cols = ['age', 'serious_dlqin2yrs']
age_means = df[cols].groupby("age").mean()
age_means.plot()

####Bin `age` into 14 different groups. Then make a frequency table that shows the number of customers that were/were not delinquent for each bin.
*HINT: You might want to have larger bins near the min/max values to account for outliers.*

mybins = [0] + range(20, 80, 5) + [120]
df['age_bucket'] = pd.cut(df.age, bins=mybins)
pd.value_counts(df['age_bin'])

####Using the age bins, calculate the percent of customers that were delinquent for each bucket

df[["age_bucket", "serious_dlqin2yrs"]].groupby("age_bucket").mean()

df[["age_bucket", "serious_dlqin2yrs"]].groupby("age_bucket").mean().plot()

labels, levels = pd.factorize(df.age_bucket)
df.age_bucket = labels

####Write something that buckets debt_ratio into 4 (nearly) equally sized groups.
*Hint: use the `quantile` method for Series*

bins = []

for q in [0.2, 0.4, 0.6, 0.8, 1.0]:
    bins.append(df.debt_ratio.quantile(q))

debt_ratio_binned = pd.cut(df.debt_ratio, bins=bins)
debt_ratio_binned
print pd.value_counts(debt_ratio_binned)

###Scaling Features
Some algorithms will work better if your data is centered around 0. The `StandardScaler` module in `scikit-learn` makes it very easy to quickly scale columns in your data frame.

from sklearn.preprocessing import StandardScaler

df['monthly_income_scaled'] = StandardScaler().fit_transform(df.monthly_income)

print df.monthly_income_scaled.describe()
print
print "Mean at 0?", round(df.monthly_income_scaled.mean(), 10)==0

pl.hist(df.monthly_income_scaled)

##Redo our feature importance

features = np.array(['revolving_utilization_of_unsecured_lines',
                     'age', 'number_of_time30-59_days_past_due_not_worse',
                     'debt_ratio', 'monthly_income','number_of_open_credit_lines_and_loans', 
                     'number_of_times90_days_late', 'number_real_estate_loans_or_lines',
                     'number_of_time60-89_days_past_due_not_worse', 'number_of_dependents',
                     'income_bins', 'age_bucket', 'monthly_income_scaled'])

clf = RandomForestClassifier(compute_importances=True)
clf.fit(df[features], df['serious_dlqin2yrs'])

importances = clf.feature_importances_
sorted_idx = np.argsort(importances)

padding = np.arange(len(features)) + 0.5
pl.barh(padding, importances[sorted_idx], align='center')
pl.yticks(padding, features[sorted_idx])
pl.xlabel("Relative Importance")
pl.title("Variable Importance")
pl.show()

best_features = features[sorted_idx][::-1]
best_features

##Test out your own features
Partner with the person sitting next to you and see if you can come up with some new features that outperform the basic variables.

- brainstorm new features
- create them using pandas
- test their effectiveness w/ scikit-learn

Things you might try:

- Try using some of the scikit-learn [preprecessing functions](http://scikit-learn.org/stable/modules/preprocessing.html#preprocessing)
- Look at `pd.get_dummies`
- Combine features together (total times late, log(monthly_income), age^2)

##We just did the following

- Ranked feature importance using RandomForest
- Created some of our own features using `pandas` and `scikit-learn`
- Scaled data using `scikit-learn`
