==> 7 - Fitting and Evaluating Your Model.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- Intro to scikit-learn
- Grouping and aggregating data
- Feature selection
- <p style="color: red">Fitting and evaluating a model</p>
- Deploying your work

#In this workbook you will
- Create your own classifier for predicting delinquent customers
- Build basic reports to help interpret the effectiveness of your model
- Convert the results of your model to a credit score

import pandas as pd
import numpy as np
import pylab as pl

train = pd.read_csv("./data/credit-data-trainingset.csv")
test = pd.read_csv("./data/credit-data-testset.csv")
test.head()

##Import some classifiers

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.svm import SVC

##Fitting your model

features = ['revolving_utilization_of_unsecured_lines', 'debt_ratio',
            'monthly_income', 'age', 'number_of_times90_days_late']

clf = KNeighborsClassifier(n_neighbors=13, warn_on_equidistant=False)

clf.fit(train[features], train.serious_dlqin2yrs)

###Generating Predictions

#classes (returns an array)
clf.predict(test[features])

#probabilities (returns a numpy array)
clf.predict_proba(test[features])

###Plot a histogram of the probabilities

probs = clf.predict_proba(test[features])
prob_true = probs[::,1]
pl.hist(prob_true)

##[Evaluating your model](http://scikit-learn.org/stable/modules/model_evaluation.html#classification-metrics)

from sklearn.metrics import roc_curve, auc, classification_report, confusion_matrix

preds = clf.predict_proba(test[features])
preds

##Classification reports

###What does this tell you?

confusion_matrix(test['serious_dlqin2yrs'], clf.predict(test[features]))

print classification_report(test['serious_dlqin2yrs'], clf.predict(test[features]), labels=[0, 1])

###Write your own version of confusion_matrix using `pandas`. Be sure to label the rows/columns.
*HINT: use crosstab*

pd.crosstab(test['serious_dlqin2yrs'], clf.predict(test[features]), rownames=["Actual"], colnames=["Predicted"])

###The ROC Curve
To evaluate our classifier, we're going to use an [ROC Curve](http://en.wikipedia.org/wiki/Receiver_operating_characteristic). ROC Curves are great for evaluating binary (0, 1) classification models. A ROC Curve plots the False Postive Rate (fpr) vs. the True Positive Rate (tpr) for a classifier.

See example in [scikit-learn docs](http://scikit-learn.org/stable/auto_examples/plot_roc.html#example-plot-roc-py).

def plot_roc(name, probs):
    fpr, tpr, thresholds = roc_curve(test['serious_dlqin2yrs'], probs)
    roc_auc = auc(fpr, tpr)
    pl.clf()
    pl.plot(fpr, tpr, label='ROC curve (area = %0.2f)' % roc_auc)
    pl.plot([0, 1], [0, 1], 'k--')
    pl.xlim([0.0, 1.05])
    pl.ylim([0.0, 1.05])
    pl.xlabel('False Positive Rate')
    pl.ylabel('True Positive Rate')
    pl.title(name)
    pl.legend(loc="lower right")
    pl.show()
plot_roc("Perfect Classifier", test['serious_dlqin2yrs'])
plot_roc("Guessing", np.random.uniform(0, 1, len(test['serious_dlqin2yrs'])))

#[::,1] selects the 2nd column of the numpy array
plot_roc("KNN", preds[::,1])

clf = RandomForestClassifier()
clf.fit(train[features], train.serious_dlqin2yrs)

probs = clf.predict_proba(test[features])[::,1]
plot_roc("RandomForest", probs)

###Build your own classifier
Pick a different algorithm and a new set of features. Can you beat the 0.74 AUC?

train.head()

features = ['revolving_utilization_of_unsecured_lines', 'debt_ratio',
            'number_of_times90_days_late', 'number_real_estate_loans_or_lines']
clf = GradientBoostingClassifier()
clf.fit(train[features], train.serious_dlqin2yrs)
probs = clf.predict_proba(test[features])[::,1]
plot_roc("Your Classifier", probs)

###Converting to credit score
We're going to take the P(delinquent) outputted by the model and convert it to a FICO style score. We calculate the log odds which we then convert into 'points'. in this case, a increase/decrease in 40 points (arbritrary) means a person's riskness has halved/doubled--40/log(2). We're starting with a base score of 340 (arbitrary).

probs
odds = (1 - probs) / probs
score = np.log(odds)*(40/np.log(2)) + 340
pl.hist(score)

def convert_prob_to_score(p):
    """
    takes a probability and converts it to a score
    Example:
        convert_prob_to_score(0.1)
        466
    """
    odds = (1 - p) / p
    scores = np.log(odds)*(40/np.log(2)) + 340
    return scores.astype(np.int)

convert_prob_to_score(probs)

##We just did the following

- Trained a credit classifier
- Evaluated the results by both plotting data and generating reports
- Converted the model into a credit score
